+++ include/nctl/LinearAllocator.h
@@ -28,8 +28,8 @@ class DLL_PUBLIC LinearAllocator : public IAllocator
 	LinearAllocator(const LinearAllocator &) = delete;
 	LinearAllocator &operator=(const LinearAllocator &) = delete;
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
