+++ include/nctl/MallocAllocator.h
@@ -18,8 +18,8 @@ class DLL_PUBLIC MallocAllocator : public IAllocator
 	MallocAllocator(const MallocAllocator &) = delete;
 	MallocAllocator &operator=(const MallocAllocator &) = delete;
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
