+++ include/nctl/PoolAllocator.h
@@ -16,20 +16,20 @@ class DLL_PUBLIC PoolAllocator : public IAllocator
 	    : PoolAllocator("Pool", elementSize, DefaultAlignment, size, base) {}
 	PoolAllocator(const char *name, size_t elementSize, size_t size, void *base)
 	    : PoolAllocator(name, elementSize, DefaultAlignment, size, base) {}
-	PoolAllocator(size_t elementSize, uint8_t elementAlignment, size_t size, void *base)
+	PoolAllocator(size_t elementSize, size_t elementAlignment, size_t size, void *base)
 	    : PoolAllocator("Pool", elementSize, elementAlignment, size, base) {}
-	PoolAllocator(const char *name, size_t elementSize, uint8_t elementAlignment, size_t size, void *base);
+	PoolAllocator(const char *name, size_t elementSize, size_t elementAlignment, size_t size, void *base);
 	~PoolAllocator();
 
 	inline void init(size_t elementSize, size_t size, void *base) { init(elementSize, DefaultAlignment, size, base); }
-	void init(size_t elementSize, uint8_t elementAlignment, size_t size, void *base);
+	void init(size_t elementSize, size_t elementAlignment, size_t size, void *base);
 	inline size_t elementSize() const { return elementSize_; }
-	inline uint8_t elementAlignment() const { return elementAlignment_; }
+	inline size_t elementAlignment() const { return elementAlignment_; }
 	void **freeList() const { return freeList_; }
 
   private:
 	size_t elementSize_;
-	uint8_t elementAlignment_;
+	size_t elementAlignment_;
 	void **freeList_;
 
 	PoolAllocator(const PoolAllocator &) = delete;
@@ -37,8 +37,8 @@ class DLL_PUBLIC PoolAllocator : public IAllocator
 
 	void internalInit();
 
-	static void *allocateImpl(IAllocator *allocator, size_t size, uint8_t alignment);
-	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, uint8_t alignment, size_t &oldSize);
+	static void *allocateImpl(IAllocator *allocator, size_t size, size_t alignment);
+	static void *reallocateImpl(IAllocator *allocator, void *ptr, size_t size, size_t alignment, size_t &oldSize);
 	static void deallocateImpl(IAllocator *allocator, void *ptr);
 };
